+++ crates/fuel-core/src/database/transactions.rs
@@ -5,7 +5,6 @@ use crate::database::{
     DatabaseError,
     Result as DatabaseResult,
 };
-use fuel_core_database::vm_database::TxIdOwnerRecorder;
 use fuel_core_storage::{
     iter::IterDirection,
     tables::Transactions,
@@ -27,6 +26,7 @@ use std::{
     mem::size_of,
     ops::Deref,
 };
+use fuel_core_storage::database::TxIdOwnerRecorder;
 
 impl DatabaseColumn for Transactions {
     fn column() -> Column {
