+++ crates/services/executor/src/refs.rs
@@ -1,5 +1,4 @@
 mod contract;
-mod database;
 
 pub use contract::{
     ContractRef,
