+++ crates/storage/Cargo.toml
@@ -22,6 +22,7 @@ derive_more = { workspace = true }
 fuel-core-types = { workspace = true, default-features = false }
 fuel-vm-private = { workspace = true, default-features = false }
 mockall = { workspace = true, optional = true }
+serde = { workspace = true, features = ["derive"] }
 
 [features]
 test-helpers = ["dep:mockall"]
